Fixed 2 issues causing an incorrect FLWOR result.
Issues where not caused by the FLWOR query, but in one case, there was an uncommitted active component written to the database at the time a new component version was created. In the second case there was a concurrent modification side-effect of a stream. Fixed both with respect to the ComponentIsActive clause.